+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [Forge-1.16.1-2.4.0] - 2020-09-21
+* Add option to cut trees with mixed log blocks
+
 ## [Forge-1.16.1-2.3.1] - 2020-09-09
 * Check world type before scheduling leaf breaking (#23)
 * Do not take into account "minimumLeavesAroundRequired" when using shift down mode